refactor: move CreateDirectory function to services package and update references
