+++ cli/cmd/moling_test.go
@@ -22,12 +22,12 @@ import (
 )
 
 func TestNewMLServer(t *testing.T) {
-	err := CreateDirectory(mlConfig.BasePath)
+	err := services.CreateDirectory(mlConfig.BasePath)
 	if err != nil {
 		t.Errorf("Failed to create base directory: %v", err)
 	}
 	for _, dirName := range mlDirectories {
-		err = CreateDirectory(filepath.Join(mlConfig.BasePath, dirName))
+		err = services.CreateDirectory(filepath.Join(mlConfig.BasePath, dirName))
 		if err != nil {
 			t.Errorf("Failed to create directory %s: %v", dirName, err)
 		}
